// app.client/src/api/models/ProjectListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProjectListVM = {
    projectId?: string | null;
    name?: string | null;
    description?: string | null;
    startDate?: string | null;
    endDate?: string | null;
    status?: string | null;
    isActive?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    statusName?: string | null;
};