// app.client/src/api/models/FolderIntermediateListByFolderVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FolderIntermediateListByFolderVM = {
    folderIntermediateId?: string | null;
    folderId?: string | null;
    category?: string | null;
    categoryId?: string | null;
    isActive?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};